refactor(analysis-types): split import DTOs
